os: set MTU value for guest OS eth0 interfaces
Mismatched MTU values during packet transmission may result in packet
fragmentation, causing eBPF code to mishandle them and generating
unexpected network traffic errors.

So far we have found 1460 to be a somewhat lowest-common-denominator for
regular network traffic across the Internet. This value may be adjusted
in the future as we collect more data.

Signed-off-by: Luiz Aoqui <luizaoqui@loopholelabs.io>